Split cnn_learner and cnn_classifier into seperate files
