+++ medinify/classifiers/__init__.py
@@ -1,6 +1,6 @@
 
 from .cnn_learner import CNNLearner
-from .cnn_learner import ClassificationNetwork
+from .cnn_learner import ClassificationCNN
 from .classifier import Classifier
 from .model import Model
 from .utils import print_evaluation_metrics
